Add files via upload
For my last semester In my Advanced Software Development class we were placed into two different scrum teams made up of five people each. My scrum team was instructed to create a  product that is a graphic user interface that is connected to a database that had to be created from scratch. We had 10 weeks to complete this project during 4 different sprint periods. Our product was called Grade Checker and it would be written in C# through a Windows Form Application. This product would be used by students to check their grades for certain classes and see their grade for that class and Semester GPA and Cumulative GPA.  their were 3 different data grid view tables that showed different info. The GUI also has a search/filter function for each of the three data grid views. Look at the attached Grade Checker GUI Design.jpg file to see the design of the GUI  My role for this project was to create the GUI from scratch and then connect the database to windows form app/GUI.